+++ src/index.tsx
@@ -1,11 +1,11 @@
 import { Suspense } from 'react'
 import { Provider } from './Provider'
-import { context } from './setup'
 
 import type { ProviderProps } from './Provider'
 
 export * from './Debug'
 export * from './hooks'
+export * from './setup'
 
 function Physics(props: ProviderProps) {
   return (
@@ -15,4 +15,4 @@ function Physics(props: ProviderProps) {
   )
 }
 
-export { Physics, context }
+export { Physics }